Added dashboard homepage and moved route links to separate links.py file
