+++ website/models.py
@@ -15,4 +15,5 @@ class User(db.Model, UserMixin):
     email = db.Column(db.String(150), unique=True)
     password = db.Column(db.String(150))
     first_name = db.Column(db.String(150))
-    notes = db.relationship('Note')
+    income = db.Column(db.Integer, default=-1)
+
