+++ website/templates/apply.html
@@ -1,0 +1,5 @@
+{% extends "base.html" %} {% block title %}User Dashboard{% endblock %} {% block content %}
+
+<h1 align="center">Credit Card Application</h1>
+
+{% endblock %}
