+++ website/templates/base.html
@@ -32,7 +32,7 @@
         <div class="navbar-nav ms-auto">
           {% if user.is_authenticated %}
           <a class="nav-item nav-link" id="home" href="/">Home</a>
-          <a class="nav-item nav-link" id="user-info" href="/user-info">Personal Info</a>
+          <a class="nav-item nav-link" id="dashboard" href="/dashboard">Dashboard</a>
           <a class="nav-item nav-link" id="logout" href="/logout">Logout</a>
           {% else %}
           <a class="nav-item nav-link" id="login" href="/login">Login</a>
