+++ website/templates/spending_habits.html
@@ -1,0 +1,5 @@
+{% extends "base.html" %} {% block title %}User Dashboard{% endblock %} {% block content %}
+
+<h1 align="center">{{ user.first_name }}'s Spending Habits</h1>
+
+{% endblock %}
